+++ src/test/java/Costiil2/EstrellaTest.java
@@ -1,4 +1,4 @@
-package ch.heigvd.dai.chill.domain.Costiil2;
+package Costiil2;
 
 import ch.heigvd.dai.chill.domain.Bartender;
 import ch.heigvd.dai.chill.domain.Costiil2.Estrella;
