+++ src/test/java/Costiil2/StellaArtoisTest.java
@@ -1,7 +1,7 @@
-package ch.heigvd.dai.chill.domain.Costiil2;
+package Costiil2;
 
 import ch.heigvd.dai.chill.domain.Bartender;
-import ch.heigvd.dai.chill.domain.Costiil2.Boxer;
+import ch.heigvd.dai.chill.domain.Costiil2.StellaArtois;
 import ch.heigvd.dai.chill.protocol.OrderRequest;
 import ch.heigvd.dai.chill.protocol.OrderResponse;
 import org.junit.jupiter.api.Test;
@@ -10,22 +10,22 @@ import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class BoxerTest {
+class StellaArtoisTest {
 
   @Test
   void thePriceAndNameForPunkIPAShouldBeCorrect() {
-    Boxer beer = new Boxer();
-    assertEquals(beer.getName(), Boxer.NAME);
-    assertEquals(beer.getPrice(), Boxer.PRICE);
+    StellaArtois beer = new StellaArtois();
+    assertEquals(beer.getName(), StellaArtois.NAME);
+    assertEquals(beer.getPrice(), StellaArtois.PRICE);
   }
 
   @Test
-  void aBartenderShouldAcceptAnOrderForPunkIPA() {
+  void aBartenderShouldAcceptAnOrderForStellaArtois() {
     Bartender jane = new Bartender();
-    String productName = "ch.heigvd.dai.chill.domain.Costiil2.Boxer";
+    String productName = "ch.heigvd.dai.chill.domain.Costiil2.StellaArtois";
     OrderRequest request = new OrderRequest(3, productName);
     OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = Boxer.PRICE.multiply(new BigDecimal(3));
+    BigDecimal expectedTotalPrice = StellaArtois.PRICE.multiply(new BigDecimal(3));
     assertEquals(expectedTotalPrice, response.getTotalPrice());
   }
